made userid an int, which ended up touching a lot
Former-commit-id: de3d509ad3538ecb419fc30f364e37090a5af668
Former-commit-id: b3aabcf7de82e7221f1b7e9b8a7f2db0d9ac84dd [formerly 64b30412a7a867a4996a51baeeb82ce9eb33b40d]
Former-commit-id: 61d7dd27f48aa5762b2a4cf03e2044a26e11b743

+++ src/mitll/langtest/server/ScoreServlet.java
@@ -488,7 +488,7 @@ public class ScoreServlet extends DatabaseServlet {
   private void gotLogEvent(HttpServletRequest request, String device, JSONObject jsonObject) {
     String user = request.getHeader(USER);
 
-    long userid = userManagement.getUserFromParam2(user);
+    int userid = userManagement.getUserFromParam2(user);
     if (getUser(userid) == null) {
       jsonObject.put(ERROR, "unknown user " + userid);
     } else {
@@ -507,11 +507,11 @@ public class ScoreServlet extends DatabaseServlet {
     }
   }
 
-  private User getUser(long userid) {
+  private User getUser(int userid) {
     return db.getUserDAO().getUserWhere(userid);
   }
 
-  private String getUserID(long userid) {
+  private String getUserID(int userid) {
     User userWhere = db.getUserDAO().getUserWhere(userid);
     if (userWhere == null) logger.error("huh? can't find user by " + userid);
     return userWhere == null ? "" + userid : userWhere.getUserID();
